bpf: replaced basic monitoring with bpf, other parts not changed yet
Capmon now uses the bpf fexit probe on ns_capable. This attaches to the
return part of a function, allowing access to both call arguments and
return value.

Other kprobes such as cap_capable and capable_wrt_inode_uid_gid are not
available yet. They are still activated as of this commit, but not used
when monitoring.

Filter and summary mode are adapted to bpf mode.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>